+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -244,16 +244,14 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     return manifest;
   }
 
-  async getServerManifestAsync({
-    environment,
-  }: Pick<ExpoMetroOptions, 'environment'> = {}): Promise<{
+  async getServerManifestAsync(): Promise<{
     serverManifest: ExpoRouterServerManifestV1;
   }> {
     // NOTE: This could probably be folded back into `renderStaticContent` when expo-asset and font support RSC.
     const { getBuildTimeServerManifestAsync } = await this.ssrLoadModule<
       typeof import('expo-router/build/static/getServerManifest')
     >('expo-router/build/static/getServerManifest.js', {
-      environment: environment ?? (this.isReactServerComponentsEnabled ? 'react-server' : 'node'),
+      environment: 'node',
     });
 
     return {